ajout de fonctionnalites de sauvegarde : - tache rake pour l'exportation des donnees sur les variable : export_vdf - tache rake pour l'importation : import_vdf - script de dump de la base postgre
resolution ticket #22

+++ carte_curseur_v1/public/javascripts/initdiscretisation.js
@@ -524,13 +524,13 @@ function verifierCoherenceIntervals(tabInter){
 	nbCl = tabInter.length;
 	
 	for (var i=0; i<(nbCl-1); i++){
-		if (parseInt(tabInter[i]) > parseInt(tabInter[i+1])){
+		if (parseFloat(tabInter[i]) > parseFloat(tabInter[i+1])){
 			err += 1 ;
 		}
 		if (tabInter[i].length < 1){
 			err += 1 ;
 		}	
-		if (parseInt(tabInter[i]) == parseInt(tabInter[i+1])){
+		if (parseFloat(tabInter[i]) == parseFloat(tabInter[i+1])){
 			err += 1 ;
 		}
 	}
